Extract report card component
